+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - added `charSpacing` and `lineSpacing` in `TextCompOpt` and `DrawTextOpt`
 - added optional `transitions` argument in `state()` to define allowed transitions
 - added `StateComp#onStateTransition` to register event for specific transitions
-- added `outView()` component
+- added `outOfView()` component
 - deprecated `cleanup(number?: time)` in favor of `cleanup(opt?: CleanupOpt)`
 
 ### v2000.1.6